+++ src/refactoring/hooks/useAccordion.ts
@@ -1,0 +1,22 @@
+import { useState } from "react";
+
+export const useAccordion = () => {
+  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
+
+  const toggleProductAccordion = (productId: string) => {
+    setOpenProductIds((prev) => {
+      const newSet = new Set(prev);
+      if (newSet.has(productId)) {
+        newSet.delete(productId);
+      } else {
+        newSet.add(productId);
+      }
+      return newSet;
+    });
+  };
+
+  return {
+    openProductIds,
+    toggleProductAccordion,
+  };
+};
